# Optimize Leaflet map markers