Part of https://github.com/ParkingReformNetwork/reform-map/issues/593. DOM operations are expensive with Leaflet, so only add and remove entries when it's absolutely necessary.

Before:

[Debug] mapMarkers: 21.648ms (localhost, line 38737)
[Debug] mapMarkers: 11.689ms (localhost, line 38737)
[Debug] mapMarkers: 310.842ms (localhost, line 38737)
[Debug] mapMarkers: 28.006ms (localhost, line 38737)
[Debug] mapMarkers: 320.170ms (localhost, line 38737)
[Debug] mapMarkers: 23.577ms (localhost, line 38737)
[Debug] mapMarkers: 304.805ms (localhost, line 38737)
[Debug] mapMarkers: 19.426ms (localhost, line 38737)

After:

[Debug] mapMarkers: 4.816ms (localhost, line 38739)
[Debug] mapMarkers: 0.016ms (localhost, line 38739)
[Debug] mapMarkers: 64.560ms (localhost, line 38739)
[Debug] mapMarkers: 15.234ms (localhost, line 38739)
[Debug] mapMarkers: 49.887ms (localhost, line 38739)
[Debug] mapMarkers: 11.549ms (localhost, line 38739)
[Debug] mapMarkers: 42.215ms (localhost, line 38739)